% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{BootElementRatio}
\alias{BootElementRatio}
\title{Parametric Bootstrap of ElementRatio object}
\usage{
BootElementRatio(x, hessian = !is.null(x$optim$hessian), ...)
}
\arguments{
\item{x}{the ElementRatio object}

\item{hessian}{get the hessian from the optimization}

\item{...}{additional parameters sent to \code{ControlElementRatioFit}}
}
\value{
a new ElementRatio object, with bootstrapped data
}
\description{
\code{BootElementRatio}
}

