% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ElementRatios}
\alias{coef.ElementRatios}
\title{get coefficients for an ElementRatios object}
\usage{
\method{coef}{ElementRatios}(object, type = c("output", "par", "par.long"),
  ...)
}
\arguments{
\item{object}{the ElementRatios object}

\item{type}{Either output, par, or par.long; output yields the parameters
that are most interpretable, par yields the fitted parameters, and par.long
yields the fitted parameters in long format with the confidence interval,
if available}

\item{...}{additional parameters, unused}
}
\description{
get coefficients for an ElementRatios object
}

