% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{ControlElementRatioFit}
\alias{ControlElementRatioFit}
\title{Set the control parameters for the optimization of the element ratio fit}
\usage{
ControlElementRatioFit(depth, logratio, p1.start = 0.2, p2.start = 0.5,
  c.start = diff(range(logratio))/2, x2.start = p2.start * max(depth),
  par.start, lower = c(p1 = 0, x2 = 0.05 * max(depth), c = 0, sd1 = 0.1 *
  sd(logratio), sd2 = 0.1 * sd(logratio), d = min(logratio)), upper = c(p1 =
  0.95, x2 = max(depth), c = diff(range(logratio)), sd1 = sd(range(logratio)),
  sd2 = sd(range(logratio)), d = max(logratio)), parscale = c(p1 = 1, x2 =
  max(depth), c = 1, sd1 = 1, sd2 = 1, d = 1))
}
\arguments{
\item{depth}{depth}

\item{logratio}{logratio}

\item{p1.start}{p1.start}

\item{p2.start}{p2.start}

\item{c.start}{c.start}

\item{x2.start}{x2.start}

\item{par.start}{par.start}

\item{lower}{lower}

\item{upper}{upper}

\item{parscale}{parscale}
}
\value{
data frame with the starting parameters, the lower and upper bounds, and the scaling
}
\description{
\code{ControlElementRatioFit} sets the control parameters for the optimization
of the element ratio fit
}

