% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{extractFrom}
\alias{extractFrom}
\alias{extractFrom.ElementRatio}
\alias{extractFrom.ElementRatios}
\title{Extract information from ElementRatio fits}
\usage{
extractFrom(x, ...)

\method{extractFrom}{ElementRatio}(x, what = c("data", "lines", "cis"),
  log = TRUE, addnames = FALSE, ...)

\method{extractFrom}{ElementRatios}(x, what = c("data", "lines", "cis"),
  log = TRUE, addnames = TRUE, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}

\item{what}{either data, lines, or cis}

\item{log}{should response be on logratio scale?}

\item{addnames}{should the mobile and immobile names be added to the output data frame}
}
\value{
a data frame with the desired information
}
\description{
Extract information from ElementRatio fits
}
\section{Methods (by class)}{
\itemize{
\item \code{ElementRatio}: for ElementRatio

\item \code{ElementRatios}: for ElementRatios
}}

