% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{FitConstantRatios}
\alias{FitConstantRatios}
\title{Refit to have the same ratio fit over several immobile elements}
\usage{
FitConstantRatios(f, controlfun = ControlConstantRatioFit, verbose = TRUE)
}
\arguments{
\item{f}{ElementRatios object}

\item{controlfun}{function to get control parameters for each mobile element fit}

\item{verbose}{set verbosity}
}
\value{
an ElementRatios object, which is a list of lists of ElementRatio objects
}
\description{
\code{FitConstantRatios}
}

