% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{FitElementRatio}
\alias{FitElementRatio}
\title{Fit the element ratio for a single pair}
\usage{
FitElementRatio(mobile, immobile, depth, data, min.mobile, profile = "d",
  loglinear = FALSE, verbose = TRUE, hessian = FALSE, ...)
}
\arguments{
\item{mobile}{the name of the mobile variable}

\item{immobile}{the name of the immobile variable}

\item{depth}{the name of the depth variable}

\item{data}{the data set where these variables are found}

\item{min.mobile}{will set zero mobile values to this value (default is the smallest value)}

\item{profile}{variables to profile over automatically}

\item{loglinear}{allows the linear portion to be linear on the log(ratio) scale instead of the ratio scale}

\item{verbose}{set verbosity}

\item{hessian}{get the hessian from the optimization}

\item{...}{additional parameters sent to \code{ControlElementRatioFit}}
}
\value{
an ElementRatio object, with the output from the optimization and the optimal parameters, presented both in the
}
\description{
\code{FitElementRatio} creates the mobile/immobile ratio and fits it to the depth.
}

