% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{FitElementRatios}
\alias{FitElementRatios}
\title{Fit a bunch of element ratios all at once}
\usage{
FitElementRatios(mobile, immobile, depth, data, ...)
}
\arguments{
\item{mobile}{vector of mobile variable names}

\item{immobile}{vector of immobile variable names}

\item{depth}{name of the depth variable}

\item{data}{data frame where these variables are found}

\item{...}{additional parameters sent to \code{ControlElementRatioFit}}
}
\value{
a list of lists of ElementRatio objects
}
\description{
\code{FitElementRatios} runs \code{FitElementRatio} on several elements
}

