% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{FitTau}
\alias{FitTau}
\title{Fit the element ratio for a single pair}
\usage{
FitTau(mobile, immobile, depth, data, cutoff, log = FALSE, verbose = TRUE)
}
\arguments{
\item{mobile}{the name of the mobile variable}

\item{immobile}{the name of the immobile variable}

\item{depth}{the name of the depth variable}

\item{data}{the data set where these variables are found}

\item{cutoff}{how deep to assume constant}

\item{log}{compute on log10 scale or ratio scale}

\item{verbose}{set verbosity}
}
\value{
list with data frame (with depth, ratio, tau) and the mean depth beyond the cutoff
}
\description{
\code{FitElementRatio} creates the mobile/immobile ratio and fits it to the depth.
}

