% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{FitTaus}
\alias{FitTaus}
\title{Fit a bunch of Taus all at once}
\usage{
FitTaus(mobile, immobile, depth, data, cutoff)
}
\arguments{
\item{mobile}{vector of mobile variable names}

\item{immobile}{vector of immobile variable names}

\item{depth}{name of the depth variable}

\item{data}{data frame where these variables are found}

\item{cutoff}{how deep to assume constant, a vector of same length as immobile, or a scalar to use the same depth for all}
}
\value{
an Taus object, which is a list of lists Tau objects
}
\description{
\code{FitTaus} runs \code{FitTau} on several elements
}

