% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{plot.Tau}
\alias{plot.Tau}
\title{plot a Tau object}
\usage{
\method{plot}{Tau}(x, xlim = NULL, ylim = NULL, main = paste0(x$mobile,
  "/", x$immobile), ...)
}
\arguments{
\item{x}{object}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{main}{main title of plot}

\item{...}{additional parameters sent to plot}
}
\description{
plot a Tau object
}

