% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{plot.Taus}
\alias{plot.Taus}
\title{plot a Taus object}
\usage{
\method{plot}{Taus}(x, scales = c("same", "free"), ...)
}
\arguments{
\item{x}{the object}

\item{scales}{set axis width; sliced will use same width for all plots, but allow for shifting as needed, free will set each plot differently}

\item{...}{additional parameters sent to the individual plots}
}
\description{
plot a Taus object
}

