% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ElementRatio}
\alias{print.ElementRatio}
\title{print an ElementRatio object}
\usage{
\method{print}{ElementRatio}(x, digits = 3, ...)
}
\arguments{
\item{x}{the object}

\item{digits}{number of decimal places to show}

\item{...}{unused}
}
\description{
print an ElementRatio object
}

