% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{print.Tau}
\alias{print.Tau}
\title{print a Tau object}
\usage{
\method{print}{Tau}(x, digits = 3, ...)
}
\arguments{
\item{x}{the object}

\item{digits}{how many digits to round output to}

\item{...}{unused}
}
\description{
print basic information about a Tau object
}

