% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.ElementRatio}
\alias{profile.ElementRatio}
\title{profile an ElementRatio object}
\usage{
\method{profile}{ElementRatio}(fitted, variable, alpha = 0.05,
  tolerance = 0.1, maxiter = 10, ...)
}
\arguments{
\item{fitted}{the ElementRatio object}

\item{variable}{the variable to profile over}

\item{alpha}{the desired confidence, in terms of (1-alpha) percent}

\item{tolerance}{how tight to profile it}

\item{maxiter}{maximum number of iterations}

\item{...}{unused}
}
\value{
the original object with elements profile and confint added
}
\description{
\code{profile.ElementRatio} profiles an ElementRatio object to get a
confidence interval for a desired parameter
}

