% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.ElementRatio}
\alias{residuals.ElementRatio}
\title{Extract model residuals for an ElementRatio object}
\usage{
\method{residuals}{ElementRatio}(object, type = c("response", "pearson"), ...)
}
\arguments{
\item{object}{the object}

\item{type}{either response or pearson residuals}

\item{...}{other arguments.}
}
\description{
get residuals for an ElementRatio object
}

