% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{welldata}
\alias{welldata}
\title{Example data set}
\format{\describe{
\item{depth.top}{Depth at the top of the sample}
\item{depth.bottom}{Depth at the bottom of the sample}
\item{...}{Concentrations of various elements}
}}
\usage{
welldata
}
\description{
A data set containing concentrations of several elements over a range of depths.
}
\keyword{datasets}

