% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{BootElementRatio}
\alias{BootElementRatio}
\title{Parametric Bootstrap of ElementRatio object}
\usage{
BootElementRatio(x)
}
\arguments{
\item{x}{the ElementRatio object}
}
\value{
a new ElementRatio object, with bootstrapped data
}
\description{
\code{BootElementRatio}
}

