% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{ControlElementRatioFit}
\alias{ControlElementRatioFit}
\title{Set the control parameters for the optimization of the element ratio fit}
\usage{
ControlElementRatioFit(depth, logratio, loglinear, p.start = 0.2,
  pd.start = 0.5, c.start = diff(range(logratio))/2, d.start = pd.start *
  max(depth), par.start, lower = c(p = 0, d = 0.05 * max(depth), c =
  -diff(range(logratio)), s1 = 0.1 * stats::sd(logratio), s2 = 0.1 *
  stats::sd(logratio), r = min(logratio)), upper = c(p = 0.95, d = max(depth),
  c = diff(range(logratio)), s1 = stats::sd(range(logratio)), s2 =
  stats::sd(range(logratio)), r = max(logratio)), parscale = c(p = 1, d =
  max(depth), c = 1, s1 = 1, s2 = 1, r = 1))
}
\arguments{
\item{depth}{depth}

\item{logratio}{logratio}

\item{loglinear}{allows the linear portion to be linear on the log(ratio) scale instead of the ratio scale}

\item{p.start}{p.start}

\item{pd.start}{pd.start Where d should start, in terms of proportion of maximum depth}

\item{c.start}{c.start}

\item{d.start}{d.start}

\item{par.start}{par.start}

\item{lower}{lower}

\item{upper}{upper}

\item{parscale}{parscale}
}
\value{
data frame with the starting parameters, the lower and upper bounds, and the scaling
}
\description{
\code{ControlElementRatioFit} sets the control parameters for the optimization
of the element ratio fit
}

