% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{FitElementRatio}
\alias{FitElementRatio}
\title{Fit the element ratio}
\usage{
FitElementRatio(mobile, immobile, depth, data, min.mobile, verbose = TRUE,
  hessian = FALSE, ...)
}
\arguments{
\item{mobile}{the name of the mobile variable}

\item{immobile}{the name of the immobile variable}

\item{depth}{the name of the depth variable}

\item{data}{the data set where these variables are found}

\item{min.mobile}{will set zero mobile values to this value (default is the smallest value)}

\item{verbose}{set verbosity}

\item{hessian}{get the hessian from the optimization}

\item{...}{additional parameters sent to \code{ControlElementRatioFit}}
}
\value{
an ElementRatio object, with the output from the optimization and the optimal parameters, presented both in the
}
\description{
\code{FitElementRatio}
}

