% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ElementRatio}
\alias{plot.ElementRatio}
\title{plot an ElementRatio object}
\usage{
\method{plot}{ElementRatio}(x, sd = 1, main = paste0(x$mobile, "/",
  x$immobile), log = FALSE, ylab = if (log) "logratio" else "ratio", ...)
}
\arguments{
\item{x}{object}

\item{sd}{how many sd out to put the lines}

\item{main}{title of plot}

\item{log}{to put on log scale or not}

\item{ylab}{y label}

\item{...}{additional parameters sent to plot}
}
\description{
plot an ElementRatio object
}

