% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ElementRatio}
\alias{plot.ElementRatio}
\title{plot an ElementRatio object}
\usage{
\method{plot}{ElementRatio}(x, morelines = NULL, sd = 1,
  main = paste0(x$mobile, "/", x$immobile), rotate = TRUE, log = TRUE,
  ci.lty = 3, range = NA, responselabel = if (log) "logratio" else
  "ratio", ...)
}
\arguments{
\item{x}{object}

\item{morelines}{also add lines from this ElementRatio object}

\item{sd}{how many sd out to put the lines}

\item{main}{title of plot}

\item{rotate}{should the plot be rotated to have depth on vertical axis}

\item{log}{to put on log scale or not}

\item{ci.lty}{line type of confidence intervals}

\item{range}{how wide the range of the axis should be; set to NA to set automatically}

\item{responselabel}{label of response axis}

\item{...}{additional parameters sent to plot}
}
\description{
plot an ElementRatio object
}

