% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ElementRatios}
\alias{plot.ElementRatios}
\title{plot an ElementRatios object}
\usage{
\method{plot}{ElementRatios}(x, morelines = NULL, ...)
}
\arguments{
\item{x}{the object}

\item{morelines}{also add lines from this ElementRatios object}

\item{...}{additional parameters sent to the individual plots}
}
\description{
plot an ElementRatios object
}

