% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ElementRatios}
\alias{plot.ElementRatios}
\title{plot an ElementRatios object}
\usage{
\method{plot}{ElementRatios}(x, morelines = NULL, scales = c("sliced",
  "free"), mar = c(2.5, 2.5, 2, 0), ...)
}
\arguments{
\item{x}{the object}

\item{morelines}{also add lines from this ElementRatios object}

\item{scales}{set axis width; sliced will use same width for all plots, but allow for shifting as needed, free will set each plot differently}

\item{...}{additional parameters sent to the individual plots}
}
\description{
plot an ElementRatios object
}

